Add support for wrapping long text
This came about as a way to deal with truncated treemap text. The default behaviour is to
cut off the string near the edge of the cell container and then add
ellispes. This new truncation function will add new lines until the text
is no longer extending beyond the container edge.

Some improvements: add new line before the start of a word to make
things more legible.